[benc/numeric-input-refactor] Rewrite NumericInput as a functional component
