+++ packages/perseus/src/widgets/numeric-input/numeric-input.tsx
@@ -1,3 +1,4 @@
+import {KhanMath} from "@khanacademy/kmath";
 import {useOnMountEffect} from "@khanacademy/wonder-blocks-core";
 import * as React from "react";
 import {
@@ -11,22 +12,63 @@ import {
 import {PerseusI18nContext} from "../../components/i18n-context";
 import SimpleKeypadInput from "../../components/simple-keypad-input";
 import {useDependencies} from "../../dependencies";
+import {getPromptJSON as _getPromptJSON} from "../../widget-ai-utils/numeric-input/prompt-utils";
 
 import InputWithExamples from "./input-with-examples";
-import {type NumericInputProps} from "./numeric-input.class";
 import styles from "./numeric-input.module.css";
 import stylesLegacy from "./numeric-input_legacy-styles";
-import {generateExamples, shouldShowExamples} from "./utils";
+import {
+    generateExamples,
+    normalizeCorrectAnswerForms,
+    shouldShowExamples,
+} from "./utils";
 
-import type {Focusable} from "../../types";
+import type {Focusable, Widget, WidgetExports, WidgetProps} from "../../types";
+import type {NumericInputPromptJSON} from "../../widget-ai-utils/numeric-input/prompt-utils";
+import type {
+    MathFormat,
+    PerseusNumericInputRubric,
+    PerseusNumericInputUserInput,
+    PerseusNumericInputWidgetOptions,
+} from "@khanacademy/perseus-core";
+import type {PropsFor} from "@khanacademy/wonder-blocks-core";
+
+type Props = WidgetProps<
+    PerseusNumericInputWidgetOptions,
+    PerseusNumericInputUserInput
+>;
+
+// The Widget-interface methods this component exposes via its ref. Every
+// member of Widget is optional, so this is `Required` to make the compiler
+// insist that each one listed here actually gets implemented.
+type WidgetHandle = Required<
+    Pick<
+        Widget,
+        | "focus"
+        | "focusInputPath"
+        | "blurInputPath"
+        | "getInputPaths"
+        | "getPromptJSON"
+        | "getSerializedState"
+    >
+>;
+
+// Assert that the PerseusNumericInputWidgetOptions parsed from JSON can be passed
+// as props to this component. This ensures that the PerseusNumericInputWidgetOptions
+// stays in sync with the prop types. The PropsFor<Component> type takes
+// defaultProps into account.
+// eslint-disable-next-line no-restricted-syntax
+0 as any as WidgetProps<
+    PerseusNumericInputWidgetOptions,
+    PerseusNumericInputUserInput
+> satisfies PropsFor<typeof NumericInput>;
 
 /**
- * The NumericInputComponent is a child component of the NumericInput class
- * component. It is responsible for rendering the UI elements of the Numeric
- * Input widget.
+ * The NumericInput widget is a numeric input field that supports a variety of
+ * answer forms, including integers, decimals, fractions, and mixed numbers.
  */
-export const NumericInputComponent = forwardRef<Focusable, NumericInputProps>(
-    function NumericInputComponent(props, ref) {
+export const NumericInput = forwardRef<WidgetHandle, Props>(
+    function NumericInput(props, ref) {
         const {analytics} = useDependencies();
         const context = useContext(PerseusI18nContext);
         const inputRef = useRef<Focusable>(null);
@@ -43,22 +85,45 @@ export const NumericInputComponent = forwardRef<Focusable, NumericInputProps>(
             });
         });
 
-        // Pass the focus and blur methods to the Numeric Input Class component
         useImperativeHandle(ref, () => ({
-            current: inputRef.current,
             focus: () => {
-                if (inputRef.current) {
-                    inputRef.current?.focus();
-                    setIsFocused(true);
-                }
+                inputRef.current?.focus();
+                setIsFocused(true);
+                return true;
             },
-            blur: () => {
-                if (inputRef.current) {
-                    inputRef.current?.blur();
-                    setIsFocused(false);
-                }
+
+            focusInputPath: () => {
+                inputRef.current?.focus();
+                setIsFocused(true);
+            },
+
+            blurInputPath: () => {
+                inputRef.current?.blur();
+                setIsFocused(false);
+            },
+
+            // The widget itself is an input, so we return a single empty list to
+            // indicate this.
+            getInputPaths: () => [[]],
+
+            getPromptJSON: (): NumericInputPromptJSON => _getPromptJSON(props),
+
+            /**
+             * @deprecated and likely very broken API
+             * [LEMS-3185] do not trust serializedState
+             */
+            getSerializedState: () => {
+                const {userInput, labelText, answers: _, ...rest} = props;
+                return {
+                    ...rest,
+                    answerForms: [],
+                    labelText: labelText ?? "",
+                    currentValue: userInput.currentValue,
+                };
             },
         }));
+
+        const answerForms = normalizeCorrectAnswerForms(props.answers);
 
         const handleChange = (newValue: string): void => {
             props.handleUserInput({currentValue: newValue});
@@ -133,8 +198,8 @@ export const NumericInputComponent = forwardRef<Focusable, NumericInputProps>(
                 value={props.userInput.currentValue}
                 onChange={handleChange}
                 labelText={props.labelText || context.strings.yourAnswerLabel}
-                examples={generateExamples(props.answerForms, context.strings)}
-                shouldShowExamples={shouldShowExamples(props.answerForms)}
+                examples={generateExamples(answerForms, context.strings)}
+                shouldShowExamples={shouldShowExamples(answerForms)}
                 onFocus={handleFocus}
                 onBlur={handleBlur}
                 id={props.widgetId}
@@ -145,3 +210,117 @@ export const NumericInputComponent = forwardRef<Focusable, NumericInputProps>(
         );
     },
 );
+
+function getStartUserInput(): PerseusNumericInputUserInput {
+    return {currentValue: ""};
+}
+
+/**
+ * @deprecated and likely a very broken API
+ * [LEMS-3185] do not trust serializedState
+ */
+function getUserInputFromSerializedState(
+    serializedState: any,
+): PerseusNumericInputUserInput {
+    return {
+        currentValue: serializedState.currentValue,
+    };
+}
+
+export function findPrecision(value: number) {
+    for (let i = 0; i < 10; i++) {
+        // `toFixed` handily rounds a number to a given precision...
+        // ...but also turns it into a string. so `+` turns it back
+        // into a number.
+        if (value === +value.toFixed(i)) {
+            return i;
+        }
+    }
+    return 10; // don't assume there's more precision than that
+}
+
+export function findCommonFractions(value: number) {
+    const whole = Math.floor(value);
+    if (value === whole) {
+        return;
+    }
+    const decimal = value - whole;
+    const precision = findPrecision(decimal);
+    // it's brute force, but it's honest work
+    for (let num = 1; num < 100; num++) {
+        for (let denom = 2; denom < 100; denom++) {
+            if (+(num / denom).toFixed(precision) === decimal) {
+                return {num: num + whole * denom, denom};
+            }
+        }
+    }
+}
+
+function getCorrectUserInput(
+    options: PerseusNumericInputWidgetOptions,
+): PerseusNumericInputUserInput {
+    for (const answer of options.answers) {
+        if (answer.status === "correct" && answer.value != null) {
+            if (answer.answerForms?.includes("decimal")) {
+                return {currentValue: answer.value.toString()};
+            }
+            if (answer.answerForms?.includes("improper")) {
+                const frac = findCommonFractions(answer.value);
+                if (frac) {
+                    return {currentValue: `${frac.num}/${frac.denom}`};
+                }
+            }
+            if (answer.answerForms?.includes("proper")) {
+                const frac = findCommonFractions(answer.value);
+                if (frac) {
+                    const {num, denom} = frac;
+                    if (num > denom) {
+                        const whole = Math.floor(num / denom);
+                        const remainder = num - whole * denom;
+                        return {currentValue: `${whole} ${remainder}/${denom}`};
+                    } else {
+                        return {currentValue: `${num}/${denom}`};
+                    }
+                }
+            }
+            // 🤷
+            return {currentValue: answer.value.toString()};
+        }
+    }
+    return {currentValue: ""};
+}
+
+export default {
+    name: "numeric-input",
+    displayName: "Numeric input",
+    widget: NumericInput,
+    isLintable: true,
+    getCorrectUserInput,
+    getOneCorrectAnswerFromRubric(
+        rubric: PerseusNumericInputRubric,
+    ): string | null | undefined {
+        const correctAnswers = rubric.answers.filter(
+            (answer) => answer.status === "correct",
+        );
+        const answerStrings = correctAnswers.map((answer) => {
+            // Either get the first answer form or default to decimal
+            const format: MathFormat =
+                answer.answerForms && answer.answerForms[0]
+                    ? answer.answerForms[0]
+                    : "decimal";
+
+            let answerString = KhanMath.toNumericString(answer.value!, format);
+            if (answer.maxError) {
+                answerString +=
+                    " ± " + KhanMath.toNumericString(answer.maxError, format);
+            }
+            return answerString;
+        });
+        if (answerStrings.length === 0) {
+            return;
+        }
+        return answerStrings[0];
+    },
+    getStartUserInput,
+    getUserInputFromSerializedState,
+} satisfies WidgetExports<typeof NumericInput>;
